+++ src/main/java/ee/sk/smartid/rest/dao/SignatureSessionRequest.java
@@ -57,7 +57,7 @@ public class SignatureSessionRequest implements Serializable {
     private RequestProperties requestProperties;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
-    private String initialCallbackURL;
+    private String initialCallbackUrl;
 
     public String getRelyingPartyUUID() {
         return relyingPartyUUID;
@@ -127,11 +127,11 @@ public class SignatureSessionRequest implements Serializable {
         this.requestProperties = requestProperties;
     }
 
-    public String getInitialCallbackURL() {
-        return initialCallbackURL;
+    public String getInitialCallbackUrl() {
+        return initialCallbackUrl;
     }
 
-    public void setInitialCallbackURL(String initialCallbackURL) {
-        this.initialCallbackURL = initialCallbackURL;
+    public void setInitialCallbackUrl(String initialCallbackUrl) {
+        this.initialCallbackUrl = initialCallbackUrl;
     }
 }